Go back to cherry-picking, but with more detail
Trying to do a merge failed, see
https://stackoverflow.com/questions/72818085/how-to-merge-cherry-pick-a-single-commit
since this is basically impossible with Git. So, we're 100% stuck with
the cherry-pick idea.

We can, however, alter the message, so this preserves the nice "backport
pull request" message including the original author for atribution in
the commit message. It will have to be enough.